Fix send to closed channel.
Sometimes, the code will try to write stdin even after the conn has been closed and thus result in err and attempt to this err to already closed channel. This bug research had lead us to the point where we attempt to send the empty frame into closed conn. Added check for frame len so we wouldn't send the empty one.